cephfs: mtime should not update rctime
mtime is a user modifiable field and someone can alter it to reflect a modification time in the future or past.
rctime should be depended only on ctime which itself is system depended (based on how POSIX works).
By setting an mtime in the future, we render the usage of rctime redundant until the future timestamp is met.

Signed-off-by: Theofilos Mouratidis <t.mour@cern.ch>