mount: introduce new mount syntax
Old mount device (source) have the following problems:

- mounts to the same cluster but with different fsnames
  and/or creds have identical device string which can
  confuse xfstests.

- device shown in /proc/mounts is different that what
  is used to mount.

New proposed syntax is as follows:

  # mount -t ceph cephuser@<cluster-fsid>.mycephfs2=/path

The cluster-id is optional but is always passed to the kernel
via mount() syscall (by fetching it from cluster configuration
file if available).

The mount helper tries the mounting the file system with
the new device syntax, falling back to using the old
syntax if required.

Signed-off-by: Venky Shankar <vshankar@redhat.com>