mgr/rook: persist applied drive groups and fix drive_group_loop
This commit changes OSD creation to store applied drive groups in a map
and persist that map in the mon's KV store in a map. This commit also
modifies the rook orchestrator's serve loop to periodically re-apply drive
groups, and creates a module option to specify the interval between
re-applies.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>