mgr: expose rocksdb version number in mgr for use in telemetry
Capturing the RocksDB version number in Telemetry would allow us to check that users are using the appropriate RocksDB version for their Ceph cluster. For instance, if a user is working in a Pacific cluster, but their RocksDB version is meant for Nautilus, that might be a problem.

An encoded number is returned. Includes have been reordered according to these guidelines: https://google.github.io/styleguide/cppguide.html#Names_and_Order_of_Includes

Signed-off-by: Laura Flores <lflores@redhat.com>