pybind/mgr/pg_autoscaler: typo default option scale-up to scale-down
Typo: `scale-up` should be `scale-down` in Module
Option.

This typo doesn't trigger a bug because we create
a key-value of `scale-down` profile in
the function `create_initial()` in `src/mon/KVMonitor.cc`.
This will override whatever is the default option
in pg_autoscaler/module.py when we start the cluster and
the monitor gets created.

The command: `ceph osd pool set autoscale-profile <option>`
is still the primary command to change the autoscale-profiler
after the pool is created.

Fixes: https://tracker.ceph.com/issues/53203

Signed-off-by: Kamoltat <ksirivad@redhat.com>
(cherry picked from commit 97ac43822e1f54eb217701c411f49e483ce92bfd)