pybind/mgr/pg_autoscaler: typo default option scale-up to scale-down
Typo: `scale-up` should be `scale-down` in Module
Option.

This typo doesn't trigger a bug because we create
a key-value of `scale-down` profile in:
https://github.com/kamoltat/ceph/blob/master/src/mon/KVMonitor.cc#L51
which will override whatever is the default option
in pg_autoscaler/module.py when we start the cluster and
the monitor gets created.

The command: `ceph osd pool set autoscale-profile <option>`
is still the primary command to change the autoscale-profiler
after the pool is created.

Signed-off-by: Kamoltat <ksirivad@redhat.com>