mgr/cephadm: agent: don't attempt redeploy of down agents
In hindsight, this is a very bad idea. It is unlikely to have
situations in which an agent is down and a redeploy will fix it
with the exception of changes in dependencies which are already
covered elsewhere.

This also seems like it could be a source of thrashing we've seen.

Signed-off-by: Adam King <adking@redhat.com>