mgr/DaemonState: coarse locking
It turns out that having fine-grained locking for each DaemonState
leads to a lot of lock thrashing against the GIL when passing information
from C++ into the Python mgr environment. That's because PyFormatter must
hold the GIL, but we frequently need to walk large numbers of
independently locked DaemonStates that have a higher lock priority than
the Gil.  This can lead to lots of inefficient lock thrashing.

Instead, go back to a coarse lock on all daemonstate.  In practice, we
only have two primary consumers of this data: the ms_dispatch thread that
is processing updates from daemons (only one thread!) and the python
modules.  The fine-grained locking may help slightly with python vs report
contention, but it makes life very bad when python needs to enumerate all
items.

Signed-off-by: Sage Weil <sage@newdream.net>