qa/tasks/qemu: use direct-io with nbd device
For some reason, setting the qemu command line to writeback with an rbd-nbd
device may cause disk errors inside the VM.
Changing the qemu command line to cachemode=none is a workaround for this errors.
It should not effect rbd caching configuration,
as rbd cache policy is determinted by the rbd-nbd device itself,
using ceph.conf, and not using qemu command line.

Signed-off-by: Or Ozeri <oro@il.ibm.com>