librbd: reserve special metadata prefix for encryption
This commits creates extends the image metadata semantics
to account for a special type of metadata used for encryption.
This metadata is determined by a special prefix (.rbd_encryption).
This encryption metadata will be present whenever a cloned image
was formatted (i.e. unique encryption key was generated for that clone).
This definition means that a simple cloning of an image should not copy
this metadata.
When deep-copying or migrating an image, this metadata should be copied,
unless the destination image is flattened.

Signed-off-by: Or Ozeri <oro@il.ibm.com>