The spec.virtual_ip  parameter should be used when it's defined for the ingress daemon
When the ingress spec is built a virtual_ip parameter is provided in the spec and it's
expected to see the haproxy instance using the defined value.
The defined VIP is properly configured using the keepalived instance and, as per doc [1],
the ingress should be able to use this value as entrypoint for the haproxy frontend.
This patch also introduces a basic unit test for the IngressService with the purpose of
validating the config files generated for both haproxy and keepalived.

[1] https://docs.ceph.com/en/latest/cephadm/services/rgw/#high-availability-service-for-rgw

Fixes: https://tracker.ceph.com/issues/53684
Signed-off-by: Francesco Pantano <fpantano@redhat.com>
(cherry picked from commit 0688ca943e44f4c93cbfba34b796ae6c3823adf7)

Conflicts:
	src/pybind/mgr/cephadm/tests/test_services.py