BlueStore:NCB:Bug-Fix for recovery code with shared blobs
Fixes: https://tracker.ceph.com/issues/53678
Signed-off-by: gbenhano@redhat.com

NCB recovery is using a temporary BitmapAllocator to hold all found allocation.
When shared blobs are found it adds them multiple time as a bitmap can be set multiple times
with the same value, but it didn't take into account that the total allocation space is counted
and asserted against.

With this change we will disable the assertion when adding space from shared-blobs in recovery mode.