mgr/cephadm: still check agent deps if it is marked down
Right now if an agent is down, the way _check_agent works
if will return without ever going on to check the deps or
scheduled actions for that agent. This causes a few issues.
For one, if an agent is marked down and then a mgr failover
happens, even if reconfiguring the agent would put it in a working
state (e.g. changing the target ip if the active mgr has moved)
we never try it because _check_agent just returns as soon as it
sees the agent is down. Additionally, if someone purposely tried
to schedule a redeploy of a down agent for whatever reason, we
would never make good on this action.

This change allows us to still carry out the normal checks/
scheduled actions even on down agents

Fixes: https://tracker.ceph.com/issues/53723

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 09a593c8d56adabb01f2aeb0859e7885994ce687)