mon/MDSMonitor: remove redundant state change check
There are two sets of checks to state change in prepare_beacon.
Since the last commit, the first set is all covered by
`MDSMap::state_transition_valid`. So remove the first set and move the
second set to its place.

This fixes the bug that standby-replay is evicted unexpectedly.
This bug is introduced in
794d13c9ff4 (mon/MDSMonitor: reject illegal want_states from MDS)
but only reveal itself after
20509bb6c82 (MDSMonitor: handle damaged from standby-replay)

Fixes: https://tracker.ceph.com/issues/53811
Signed-off-by: 胡玮文 <huww98@outlook.com>