librbd: add an optional rados handle to ImageCtx
Migration from external clusters require creating a rados handle
for connecting to a remote ceph cluster.
The external rados handle will be created from within librbd, and
needs to be destroyed together with a respective ImageCtx.
To support this, this commit adds an optional rados handle
which is automatically destroyed together with its owning ImageCtx.

Signed-off-by: Or Ozeri <oro@il.ibm.com>