librbd/crypto: align data offset to stripe_period
Currently luks format operation sets the data offset
to be aligned with the image object size.
When using non-trivial striping, rbd objects can contain a combination
of luks header and actual encrypted data, which will pose a problem
later on for flattening a cloned image that was formatted.
Therefore, we change the data offset to align with stripe_period instead.
Furthermore, the written header is zero-padded to stripe period alignment,
to ensure that parent plaintext won't be copied-up while writing it.

Signed-off-by: Or Ozeri <oro@il.ibm.com>