qa: fill the test_file before trying to read it
When the file is empty the kernel will trigger a getattr request
to libcephfs to get the file size, and then if the read offset is
larger or equal to the file size it will just return without firing
real read to libcephfs.

This will fill the file with some contents and reopen the file to
flush the buffer and reset the file offset to 0.

Fixes: https://tracker.ceph.com/issues/53859
Signed-off-by: Xiubo Li <xiubli@redhat.com>