osd/scrub: force scrub status reporting
"pg dump" and similar commands show PG data , inc. scrubbing status
and counters. The data is published by the OSDs, but a change in
the scrub status was not enough to trigger an update.

A "force" parameter was added to "publish_stats_to_osd()", to
guarantee an update even if the main body of counters to be
published is not changed.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>