rgw/crypt - generalize putbucketencryption.
The previous logic only suported putbucketencryption to enable
sse-s3.  The protocol allows putbucketencryption to be used to
enable sse-kms by default, and the surrounding logic is now ready
to do this as well.  This commit removes the checks which stopped
this from working, so that it is now possible to use putbucketencryption
to default either sse-s3 or sse-kms on.

Signed-off-by: Marcus Watts <mwatts@redhat.com>