qa/cephadm: Added workunit test for orch cli mon
Added a test where we have 5 MONs on
5 different hosts and try to reduce
the number of MONs from 5 to 3 using
the command ``ceph orch apply mon 3``.

Also,increasing the number of MONs from
 3 to 5 using the command:
``ceph orch apply mon 5``.

Evaluating the correctness of the commands
and inspecting if there are crashes or unhealthy
cluster states.

This test was motivated by the bug:
https://tracker.ceph.com/issues/50089

Signed-off-by: Kamoltat <ksirivad@redhat.com>