osd: return appropriate error if the object is not manifest
Unlike existing object, manifest object requires valid adjacent
clones to calculate references.

If a user sends reference_chunk()---this is only for manifest
object---to the original object (not manifest object) which has
not recovered snapshots, the OSD triggers assert() because
reference_chunk() try to find adjacent clones, resulting in the assert().
This is because the original object
does not need to wait the recovery of snapshot.

So, this PR return an error if the object is not manifeset.

fixes: https://tracker.ceph.com/issues/54509

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>