neorados: Remove 'broken out' versions of calls
Previously we had versions of all the calls that took an int64_t and
optional key and namespace.

This didn't really offer much benefit and doubled the maintenance
burden for changing anything.

As such just make the IOContext constructor non-explicit, and make all
its mutators return the same IOContext so people can supply them
builder-style if they want to.

Only wrap some things in optional, when it actually makes it easier.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>