src/cephadm/cephadm: stop monitor before remove it from cluster
Motivation behind this PR is from the tracker:
https://tracker.ceph.com/issues/50089
`ceph orch apply mon <num>` command removes
the monitor from the cluster before shutting
the monitor down has created issues in the
mon/elector.cc part of ceph. The design of
the elector class assumes that we shut the monitor
down before removing them out of the cluster.
Therefore, this should be enforced everywhere
in ceph.

Correct steps to manually removing the monitor:

https://docs.ceph.com/en/latest/rados/operations/add-or-rm-mons/#removing-a-monitor-manual

In the function src/cephadm/cephadm `command_rm_daemon()`
if the daemon is a monitor, we remove it from the cluster
through calling the `CephContainer` class,
using the command: `mon remove <daemon id>`.
This is strictly after we stopped the monitor daemon using
`systemctl stop`.

In cephadm/services/cephadmservice.py,
function `pre_remove()`
we omitted the part where we remove the
monitor before destroying the daemon.

fixes: https://tracker.ceph.com/issues/50089

Signed-off-by: Kamoltat <ksirivad@redhat.com>