qa/cephfs: add check in test_acls.py
In xfstests-dev, "./check generic/abcd" doesn't end in error even when
there is no test abcd in generic. It's better to check the stdout to
verify success and print the returncode, stdout and stderr of the
command in logs so that such error can be found out by reading logs.

Signed-off-by: Rishabh Dave <ridave@redhat.com>