mgr/nfs: convert _cmd_nfs_export_apply to use Responder decorator
The "export apply" functionality is unusual in that it allows either
one or multiple nested requests to change or create an export.
The previous implementation would concatenate the results of
multiple change operations into a single string. It also would continue
making changes if one case failed, adding the error to the string
and setting a non-zero error code.

The updated version keeps the general behavior but returns structured
JSON (or other formatted data) with one entry per change request.  In
order to accomplish this and match the old behavior as closely as
possible we add an intermediate type (AppliedExportResults) that can
return both the response data (the `to_simplified` method) and track if
there was a failure mixed in with the various updates (the
`mgr_return_value` method).

Signed-off-by: John Mulligan <jmulligan@redhat.com>