mon/Monitor: deny mon metdata request if inconsistent
In handle_command (prefix == "mon metadata")
we only execute this command after election is
finished. Also, we forward this command to the
leader, so only the leader is able to deal with
this to prevent inconsistency in mon metadata,
since the leader will be the first to get
the most up-to-date metadata.

In get_mon_metadata, we check if
mon_metadata.size() > monmap->ranks.size()
and the rank of mon metadata we are trying
to retrieve is in monmap->removed_ranks
set. If so, we return -EINVAL to prevent
services from grabing outdated metadata,
since we know that the local mon_metadata
needs updating.

Signed-off-by: Kamoltat <ksirivad@redhat.com>
(cherry picked from commit 493f8b3ff83060959b3b098e805fe902dd819eab)