osd/scrub: do not start scrubbing if the PG is snap-trimming
Both 'snap-trim' and 'snaptrim-wait' PG states now prevent
scrub from starting.

Background:
A PG should not be scrubbed and trimmed concurrently. Unlike
write operations, snap trimming does not verify that a targeted
object is not in the executing scrub's chunk.
The trimmer always checked for active scrubs before starting. The
scrubber - did not. This PR fixes that omission.

Fixes: https://tracker.ceph.com/issues/49525

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>