osd: Allow timeout of mon command issued by an osd to set a config option.
The method, OSD::mon_cmd_set_config(), allows an osd to set a config option
related to mClock currently. The method waits until the mon acks the
command. This is generally not a problem. However, running tests with
failure injections could result in this method to wait indefinitely
causing other side-effects.

This method is called currently only during osd bring-up and any failures
induced during this time can cause the osd to not boot-up causing mon to
mark the osd as down. The config option being set currently during osd
bring-up is not very critical and there is a fallback mechanism in case of
failures.

Therefore, a new config option called "osd_mon_cmd_timeout" which defaults
to 0 is introduced. For normal cluster operation, the default timeout is
set to 0. But for tests, this option may be set to a non-zero value to
prevent issues as described above.

The fallback mechanism in case of timeouts/other errors saves the config
option in the config subsystem's "values" map at CONF_DEFAULT level using
set_val_default() method. This allows users to modify this at a later
point using "ceph config set" command.

Additionally, tests that have encountered problems due to this are modified
to include a non-zero value for this timeout (5 secs).

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>