osd: Add mechanism to skip waiting for an ack for mon cmd issued by an OSD
The method, OSD::mon_cmd_set_config(), sets a config option related to
mClock currently. The method waits until the mon acks the command. This
is generally not a problem. However, running tests with failure injections
during OSD boot-up could result in this method to wait indefinitely
causing other side-effects.

This method is currently called only during osd bring-up, and any failures
injected during this time can cause the osd to not boot-up due to no ack
being received from the mon causing mon to mark the osd as down. The config
option being set currently during osd bring-up is not very critical and
there is a fallback mechanism in case of failures to save it into the
in-memory "values" map.

Following are the changes:

- A new config option called "osd_mon_cmd_wait_ack" is introduced
that defaults to "True" for normal cluster operation. But for tests, this
option may be set to "False" to prevent issues as described above. For
tests, as a fail-safe, the mClock related config option is also saved in
the in-memory "values" map regardless of whether the write to the config
monitor db succeeded or not.

- The fallback mechanism in case of timeouts/other errors saves the config
option in the config subsystem's "values" map at CONF_DEFAULT level using
set_val_default() method. This allows users to modify this at a later
point using "ceph config set" command.

- Additionally, tests that have encountered problems due to this are
modified to set osd_mon_cmd_wait_ack to False.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>