osd: Implement Context based completion for mon cmd to set a config option
The method, OSD::mon_cmd_set_config(), sets a config option related to
mClock during OSD boot-up. The method used to wait on a condition variable
until the mon ack'ed the command. This was generally not a problem.
However, running tests with failure injections during OSD boot-up in some
cases resulted in this method to wait indefinitely causing other
side-effects, for e.g., the mon marking the osd as down.

Following are the changes:

The changes to mon_cmd_set_config() are generic and any osd specific
option may be set in the config monitor store using this method.

- Implement Context based completion tracking of the mon command using
  MonCmdSetConfigOnFinish which is derived from the base Context class.
  In case of failures, the finish() method is overriden to save the config
  option in the config subsystem's "values" map at CONF_DEFAULT level
  using set_val_default() method. This allows users to modify this at a
  later point using "ceph config set" cli command.

- Additionally, if requested, finish() also checks if the config option
  needs to be applied on each op shard and calls update_scheduler_config()
  on each shard. This is required for some config options related to the
  mClock scheduler.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>